Fetching Data
The homepage is now able to fetch data directly from the API GET endpoint.

For the purpose of development, a "proxy" field has been added to the package.json to workaround the Cross-Origin Resource Sharing (CORS) security guardrail. In an actual deployment, the proper API endpoints should be correctly coded in instead of using locahost.