React Context using useReducer
I've worked with react context before but only with useState and useEffect.

I've never used the useReducer. I'd a little more complicated to setup, and honestly it achieves mostly the same thing.

It's biggest difference is the fact that we can specifically set an `action_type` when we are updating the state.

So instead of directly modifying the contents of the state (say doing a state+1), we can just call an 'INCREMENT' action, and within the context definition, it will perform that process.