Handling Error Responses
There are many ways to deal with this. In this example, I'm simply checking if the fields are empty or not. There's no other data validation yet.

In fact, it's even possible to set a negative value for both numerical values, which in a deployed system, shouldn't be the case.